+++ DAL/Model/LoanModel.cs
@@ -7,11 +7,11 @@ namespace DAL.Model
   {
     public int BookFK { get; set; }
     public int UserFK { get; set; }
-    public double LoanPrice { get; set; }
+    public decimal LoanPrice { get; set; }
     public DateTime LoanDate { get; set; }
     public DateTime PlannedReturnDate { get; set; }
     public DateTime ReturnDate { get; set; }
     public int DelayDays { get; set; }
-    public double DelayPricePerDay { get; set; }
+    public decimal DelayPricePerDay { get; set; }
   }
 }
